separate unit-test and benchmark
